PlaceSummary and Country Page improvements
Country page (/places) now displays a list of countries, with
statistics (taken from the PlaceSummary). These figures are also used
to generate a colour-coded progress bar. The list of countries can be
sorted by name or total incident count.

A PlaceSummary is stored on each Country. This needs to be updates, by
invoking the GenerateSummaries handler
(/serviceops/generate-summaries). This should be done every time the
database is updated.